requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
google-generativeai==0.3.2
python-dotenv==1.0.0
eventlet==0.33.3
